Script & Algorithms Update
Final Script with name “Collect_data.py” is done.
Algorithms updated to fix some issues.
 a) to accept also —version (randInt, fixInt) parameter (they just
accept that but they do nothing different for the moment. This update
so i wont have to change again the script)
b) Avg_time now when readers are done is checked before doing the
division to avoid diving with 0 and screw the script and/or the
executions.
c) the script has comment out all the final variables for the final
executions. The UNcommented variables are just to run some examples and
confirm that everything work like a charm.
d) Before running the script please confirm that the algorithms are
compiled. i.e. run in terminal ./waf --run 'am-semifast --help’ just to
compile everything first
e) the script must be placed where it is now. you can run it from
terminal with command “python collect_data.py”